chg: dev: upgrade nl-utils and vrf-utils to enable vrf support
    In testing, we found that once setns is executed it need
to reset to old namespace or needs to be done via thread.
operationally it is convenient to run this setns via child thread
rather than making the parent process/thread to enter another
namespace. Also additional APIs are provided where the direct
process entry is required in special cases. The changes address
this issue in nl-utils.c and few new APIs for above are added
vrf-utils to optimize the code further.

Change-Id: I9090fb1aac863cbfd26177c356d0da5a64211948
Signed-off-by: krishna <elluru.kri.mohan@hpe.com>